Make product reorder level an integer, not a float
A reorder threshold is a whole number of units; fractional values (the
stray 0.07 seen in the form) are meaningless.

- Model: Product.reorder_level is now PositiveIntegerField (migration
  0014 casts existing decimals to whole numbers).
- POS/product forms (create, edit, purchase): number inputs use step=1
  min=0 and round to a non-negative integer on submit, so an old decimal
  loaded for editing can't fail the now-integer API.
- Legacy web templated views round reorder_level to an integer too.

+++ backend/catalog/models.py
@@ -94,7 +94,7 @@ class Product(TenantScopedModel):
     tax_percent = models.DecimalField(max_digits=5, decimal_places=2, default=0)
 
     track_inventory = models.BooleanField(default=True)
-    reorder_level = models.DecimalField(max_digits=12, decimal_places=2, default=5)
+    reorder_level = models.PositiveIntegerField(default=5)
     current_stock = models.DecimalField(max_digits=14, decimal_places=2, default=0, editable=False)
 
     is_active = models.BooleanField(default=True)
